backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
SQLAlchemy>=1.4
alembic>=1.10
pydantic>=1.10
python-dotenv>=1.0
pytest>=7.0
httpx>=0.23
psycopg2-binary>=2.9
supabase>=1.0
